src/services/business/complianceService.ts
import { ethers } from 'ethers';
import { ValidationError } from '@/utils/errors';
import { EventLogger } from '../audit/eventLogger';

export class ComplianceService {
  private static readonly RESTRICTED_COUNTRIES = ['XX', 'YY', 'ZZ'];
  private static readonly KYC_REQUIRED_THRESHOLD = ethers.utils.parseUnits('10000', 6);
  
  private eventLogger: EventLogger;

  constructor() {
    this.eventLogger = EventLogger.getInstance();
  }

  async validateTransaction(
    userId: string,
    amount: ethers.BigNumber,
    userKycStatus: string,
    userCountry: string
  ): Promise<void> {
    await Promise.all([
      this.checkKycRequirement(userId, amount, userKycStatus),
      this.checkCountryRestrictions(userId, userCountry),
      this.checkTransactionLimits(userId, amount)
    ]);
  }

  private async checkKycRequirement(
    userId: string,
    amount: ethers.BigNumber,
    kycStatus: string
  ): Promise<void> {
    if (amount.gte(this.KYC_REQUIRED_THRESHOLD) && kycStatus !== 'verified') {
      await this.eventLogger.logSecurityEvent(
        userId,
        'KYC_REQUIRED',
        { amount: amount.toString() }
      );
      throw new ValidationError('KYC verification required for this amount');
    }
  }

  private async checkCountryRestrictions(
    userId: string,
    country: string
  ): Promise<void> {
    if (this.RESTRICTED_COUNTRIES.includes(country)) {
      await this.eventLogger.logSecurityEvent(
        userId,
        'RESTRICTED_COUNTRY',
        { country }
      );
      throw new ValidationError('Service not available in your country');
    }
  }

  private async checkTransactionLimits(
    userId: string,
    amount: ethers.BigNumber
  ): Promise<void> {
    // Implement transaction limit checks
    // This would typically involve checking against regulatory limits
  }
}